requirements-common.txt
Django==4.1.1
djangorestframework==3.13.1
django-filter
gunicorn

# image manipulation
pillow

# per object permissions framework
#rules

# helper for nested models (not needed currently)
#drf-extensions

# needed for generation openapi schemas
pyyaml
uritemplate

# needed for live documentation of the api (swagger)
# unlike drf-yasg this library can generate openapi 3.0
drf-spectacular

# parse urls
tldextract

# jwt authentication
djangorestframework-simplejwt

# set of helpers for manage.py
# check https://github.com/django-extensions/django-extensions
django-extensions

# storage backends
django-storages
paramiko # needed for sftp
boto3 # for s3 backend

# cockroachdb
django-cockroachdb
dj_database_url
psycopg2-binary

# nicer error messages
pretty_errors

# using own fork a admin_reorder since it's not updated by parent
git+https://github.com/osuka/django-modeladmin-reorder@master